Adapt find_file for the picos environment
In addition to adjusting its behavior for the picos environment,
the function can no longer return a pointer as before.
Since dir_block is loaded into a cache, returning a pointer may
lead to the data being overwritten if the cache is reused by
other functions. Therefore, this function now returns an addr_t,
and callers must reload the corresponding data when needed.

Signed-off-by: YU-WEI,HSU <blackfire106.1231@gmail.com>