Adapt read_file for the picos environment
In addition to adjusting its workflow for the picos environment, an
extra constraint is enforced on cnt: it must be a multiple of 512,
otherwise the read request is rejected.
read_extern_buf now requires the address of a shared block memory,
and read_file copies data from the disk’s external memory into it.

Signed-off-by: YU-WEI,HSU <blackfire106.1231@gmail.com>